refactor userSlice, and fix Login problems
