+++ src/redux/configureStore.js
@@ -1,14 +1,12 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { createLogger } from 'redux-logger';
-import registerReducer from './users/registerSlice';
-import loginReducer from './users/loginSlice';
+import userReducer from './users/userSlice';
 
 const logger = createLogger();
 
 const store = configureStore({
   reducer: {
-    users: registerReducer,
-    login: loginReducer,
+    users: userReducer,
   },
 
   middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(logger),
